# openapi.yaml
openapi: 3.0.3
info:
  title: Random Image API
  description: 一个高性能的随机图片服务，支持多线程、Redis缓存、CORS支持和智能缓存管理
  version: 2.1.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
paths:
  /api:
    get:
      summary: 获取随机图片
      description: 从所有目录中随机返回一张图片
      operationId: getRandomImage
      parameters:
        - name: json
          in: query
          description: 是否返回JSON格式
          required: false
          schema:
            type: integer
            enum: [1]
      responses:
        '200':
          description: 图片文件或JSON信息
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            application/json:
              schema:
                $ref: '#/components/schemas/ImageInfo'
        '404':
          description: 没有找到图片
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: 请求过于频繁
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateLimitResponse'
...
components:
  schemas:
    ImageInfo:
      type: object
      properties:
        name:
          type: string
          description: 图片名称
        size:
          type: integer
          description: 图片大小（字节）
        uploadtime:
          type: string
          format: date-time
          description: 上传时间
        path:
          type: string
          description: 图片路径
        processingTime:
          type: integer
          description: 处理时间（毫秒）
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误类型
        message:
          type: string
          description: 错误消息
        path:
          type: string
          description: 请求路径
        processingTime:
          type: integer
          description: 处理时间（毫秒）
        timestamp:
          type: string
          description: 时间戳
    RateLimitResponse:
      type: object
      properties:
        error:
          type: string
          description: 错误类型
        message:
          type: string
          description: 错误消息
        clientIP:
          type: string
          description: 客户端IP
        limit:
          type: integer
          description: 限制次数
        window:
          type: integer
          description: 时间窗口（秒）
        banned:
          type: boolean
          description: 是否被封禁
        remainingTime:
          type: integer
          description: 剩余封禁时间（秒）
        banEndTime:
          type: string
          format: date-time
          description: 封禁结束时间
        reason:
          type: string
          description: 封禁原因
        timestamp:
          type: string
          description: 时间戳
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: query
      name: token
      description: API密钥认证
security:
  - ApiKeyAuth: []